refactor(server): isolate join pandas implem
