+++ server/weaverbird/utils/__init__.py
@@ -1,3 +1,2 @@
-from .cleaning import rename_duplicated_columns
 from .size import convert_size
 from .stopwatch import StopWatch
